commit_553 ------------------------------------------------------------------------ r553 | vboxsync | 2007-02-02 22:32:07 +0800 (Fri, 02 Feb 2007) | 2 lines
Don't disable the extended VBE mode when Windows want's to reset the graphics card. This simply did not work right since we did not reset the mode to the Windows boot mode but we set a mode <current width> x <current height> x 0 bpp which only caused an disturbing additional mode switch when chaning the display resolution. Tested with WinXP and Win2000 guests.

------------------------------------------------------------------------