commit_1890 ------------------------------------------------------------------------ r1890 | vboxsync | 2007-04-04 00:04:19 +0800 (Wed, 04 Apr 2007) | 2 lines
Attempt to fix ALSA on Linux kernels <= 2.6.17: use mmap not memalign for allocating pages. Use madvise or mprotect to separater VM area structs inside the kernel. Most SUP* functions work on cPages now (not cBytes anymore). The free functions take a cPages parameter which is used for munmap on Linux.

------------------------------------------------------------------------