commit_2962 ------------------------------------------------------------------------ r2962 | vboxsync | 2007-05-31 21:17:41 +0800 (Thu, 31 May 2007) | 2 lines
Don't use putenv if possible as setenv is more correct. The reason is that putenv expects a parameter which gets part of the environment. If the passed string is changed, the environment changes as well. Passing an automatic variable to putenv is a bug. Though it isn't wrong to pass a constant since a constant string goes never away. But gcc-4.2 warns about it.

------------------------------------------------------------------------