commit_3078 ------------------------------------------------------------------------ r3078 | vboxsync | 2007-06-08 00:18:31 +0800 (Fri, 08 Jun 2007) | 7 lines
1859: FE/Qt Mouse cursor issue:

Fix for the issue #1:
VBoxDefs::ResizeEventType event appears in case of guest video sub-system was changed for somehow even without video resolution change. In this last case the host VM window will not be resized according this event and the host mouse cursor which was unset to "default" here will not be hidden in capture state. So it is necessary to perform updateMouseClipping() for the guest resize event if the mouse cursor was captured before.

Fix for the issue #3:
When we are entering mouse integration mode we have to notify the guest system about it to make sure the guest will hide it’s cursor to let using of host's cursor for the guest mouse manipulation. This is not possible for the current API. That is why we have two cursors displayed one-above-one at the entering mouse integration mode. So we have to temporary hide host cursor and wait until guest hide it’s cursor and send a message about shape change to the host.
------------------------------------------------------------------------