commit_3767 ------------------------------------------------------------------------ r3767 | vboxsync | 2007-07-23 06:54:01 +0800 (Mon, 23 Jul 2007) | 2 lines
Reduce occurrences of the message "Spurios NAK on isa0060/serio0" from Linux guests. Only send 0xFE if we really want to release any pressed modifier, not on every focus release. The real fix is to find a better way to prevent guests to interprete the KEYPRESS/KEYRELEASE event of a modifier key in that situation.

------------------------------------------------------------------------