commit_6146 ------------------------------------------------------------------------ r6146 | vboxsync | 2007-12-19 16:58:24 +0800 (Wed, 19 Dec 2007) | 1 line
Fixed fullscreen mode on mac os x: 1. wasHidden is wrong (todo find out why), so perform show() instead of hide(). 2. must hide the mac os x stuff before setting the maximum size, or qt will (it seems) adjust it according to visible screen region. 3. the mask (setMask) rectangle needs to be adjusted or we'll get a one pixle border on the right and bottom of the screen (another qt bug?)
------------------------------------------------------------------------