commit_6715 ------------------------------------------------------------------------ r6715 | vboxsync | 2008-02-01 19:36:46 +0800 (Fri, 01 Feb 2008) | 5 lines
Mac OS X: Changed overlay icon in the bottom right of the dock icon from
the state icons to the 3D VirtualBox logo. If the vm is in pause state the
current pause image is displayed in the screen. Pause/Restore/Save states
are now indicated by an overlay image in the center of the dock icon.

------------------------------------------------------------------------