commit_6767 ------------------------------------------------------------------------ r6767 | vboxsync | 2008-02-04 01:39:44 +0800 (Mon, 04 Feb 2008) | 1 line
FE/Qt: 1) Release pressed keys even on focus loss in Paused state (for example when a warning about the ignored keyboard input is shown) to avoid stuck keys in the guest after resuming. 2) Show the keyboard input warning on key release (instead of key press) which is a kind of more expected by the user.
------------------------------------------------------------------------