commit_6804 ------------------------------------------------------------------------ r6804 | vboxsync | 2008-02-05 18:27:19 +0800 (Tue, 05 Feb 2008) | 7 lines
Fixed VMR3PowerOn regression.

The problem was that the EMT outer loop wasn't checking for state
changes in the init case (pUVM->pVM == NULL). So, if the VMR3PowerOn
request was queued and executed by the initial VMR3ReqProcessU call
we wouldn't call EMR3ExecuteVM when the VM state changed to running.

------------------------------------------------------------------------