commit_6814 ------------------------------------------------------------------------ r6814 | vboxsync | 2008-02-06 04:57:58 +0800 (Wed, 06 Feb 2008) | 4 lines
Added a new family of assertions, AssertLogRel, that's intended for cases where you should use LogRel but wish for it to kick you in strict builds. Use with care!

Also added a bunch of @todos. The 'BreakVoid' stuff must be renamed 'Break', and the current 'Break' should be called 'BreakStmt'. Also, all 'BreakVoid' macros and possibly some of the 'BreakStmt' macros are broken. Both needs addressing, but some care needs to be taken wrt branches.

------------------------------------------------------------------------