commit_6943 ------------------------------------------------------------------------ r6943 | vboxsync | 2008-02-14 20:44:45 +0800 (Thu, 14 Feb 2008) | 1 line
Additions/X11: added a script to call the xrandr application which is called by VBoxClient when it detects a display change.  This is needed as the driver has no mechanism for detecting monitor hotplugging directly.  This will have to be moved out of the Linux subtree into the common one (or duplicated in other subtrees) at some point.
------------------------------------------------------------------------