commit_7249 ------------------------------------------------------------------------ r7249 | vboxsync | 2008-03-04 01:50:55 +0800 (Tue, 04 Mar 2008) | 1 line
Use the critsect to serialize access to the pastboard so that the polling thread doesn't enter it at the same time as the service thread. Fixed the return code handling around queryNewPasteboardFormats (VERR_NOT_SUPPORTED should not propagate to Connect and Sync, even if returns codes are generally ignored by the service layer). Use pClient instead of the global (/me hates globals).
------------------------------------------------------------------------