commit_7702 ------------------------------------------------------------------------ r7702 | vboxsync | 2008-04-02 23:07:15 +0800 (Wed, 02 Apr 2008) | 1 line
Fixed a bunch of #else and #endif comments. The convention here is that on an #else we'll document what the criteria for the following body is, and for #endif the above body. So, the #else comment is the inverse of the #if and the #endif is the same as the #else, or if there isn't any #lese, the #if.
------------------------------------------------------------------------