commit_9100 ------------------------------------------------------------------------ r9100 | vboxsync | 2008-05-26 09:24:55 +0800 (Mon, 26 May 2008) | 1 line
Corrected fstsw to fnstsw. The fstsw mnemonic is assembled into fwait+fnstsw, see the intel manuals (3-432 in Vol. 2A), so the correct name for DF /7 and DF E0 is fnstsw. The disassembler did turn a 9B DF E0 sequence into two instructions like it should, so it's just the mnemonics.
------------------------------------------------------------------------