commit_9594 ------------------------------------------------------------------------ r9594 | vboxsync | 2008-06-11 17:45:26 +0800 (Wed, 11 Jun 2008) | 2 lines
If IntNet is used as a service (for lwIP) then taking it down on suspend/poweroff is way too early. The service and the users of it come later in the notification order and thus have no chance of doing proper connection shutdown. This fixes the 10 seconds suspend/poweroff delay when using iSCSI over IntNet.

------------------------------------------------------------------------