commit_10819 ------------------------------------------------------------------------ r10819 | vboxsync | 2008-07-23 01:58:54 +0800 (Wed, 23 Jul 2008) | 1 line
intnet: Use the IPRT handle table, validate against the session when looking up the handle, and do proper reference counting of the interfaces when resolving a handle. Fixed an issue in the interface destructure where anyone that could (very theorically) be sleeping on it wouldn't decrement the waiter count as we wanted them to.
------------------------------------------------------------------------