commit_11064 ------------------------------------------------------------------------ r11064 | vboxsync | 2008-08-01 22:12:37 +0800 (Fri, 01 Aug 2008) | 1 line
iprt/RTRandUxy: Improved the distribution when not using the max ranges. Changed the pseudo/cpuid fallback so it makes sure the cpuid has changed sufficiently and then only takes 3-bits from it, dropping the lower 5 since some of these are no longer reliable (5 is a bit too high, hopefully). Added tstRand for testing the distribution and limits.
------------------------------------------------------------------------