commit_14451 ------------------------------------------------------------------------ r14451 | vboxsync | 2008-11-21 19:12:56 +0800 (Fri, 21 Nov 2008) | 1 line
Config.kmk: Removed VBOX_RELEASE_BUILD, it's too much of a build and QA hazzard. In general nothing should differ between odd and even build numbers. Exception can be made on a case by case basis after discussing it with the guys doing the builds and maintaing the build system.
------------------------------------------------------------------------