commit_15505 ------------------------------------------------------------------------ r15505 | vboxsync | 2008-12-15 22:36:30 +0800 (Mon, 15 Dec 2008) | 1 line
SUPDrv,INTNet: Heads up! SupDrv version bumped. Added SUPR0ObjAddRefEx for dealing with the handle table callback which occurs while owning a spinlock. Normally SUPR0ObjAddRef[Ex] would always allocate a usage record, which means RTMemAlloc, but this is a bad idea when inside a spinlock. SUPR0ObjAddRefEx sports an additional parameter indicating whether it is allowed block or not.
------------------------------------------------------------------------