commit_16649 ------------------------------------------------------------------------ r16649 | vboxsync | 2009-02-11 02:29:02 +0800 (Wed, 11 Feb 2009) | 1 line
Main/Glue: com::Initialize() now first looks for XPCOM binaries and components in the VBOX_APP_HOME directory when this environment variable is set. Otherwise, it tries the compiled-in VirtualBox installation path as before and if it fails, checks several hard-coded paths. This is to make it possible to start 3rd-party client applications located in directories other than the VirtualBox installation directory.
------------------------------------------------------------------------