commit_16719 ------------------------------------------------------------------------ r16719 | vboxsync | 2009-02-13 08:11:24 +0800 (Fri, 13 Feb 2009) | 1 line
Config,VBoxDrv: Updated VBOXR0DRV and VBoxDrv to work with 10a261. The main trick was using the right options: -XLinker -kext and -mkernel. We used to link it as a static (MH_OBJECT) but that no longer works. Also, kextload doesn't work for development purposes, so we'll have to use kextutil instead.
------------------------------------------------------------------------