commit_17509 ------------------------------------------------------------------------ r17509 | vboxsync | 2009-03-07 09:30:23 +0800 (Sat, 07 Mar 2009) | 1 line
PGM: Moved the page pool PT flushing code in the access handler bits to where it belongs and called it pgmPoolTrackFlushGCPhys. Fixed a status code corruption bug in PGMR3PhysTlbGCPhys2Ptr (new phys). Made lazy zero page replacement code work in the new code, it's disabled by default because it frequently requires flushing the shadow page pool because the tracking code assuming the HCPhys of a PGMPAGE is unique and never shared.
------------------------------------------------------------------------