commit_18354 ------------------------------------------------------------------------ r18354 | vboxsync | 2009-03-27 06:17:20 +0800 (Fri, 27 Mar 2009) | 1 line
MMHyper: With lots of memory, we have to scale up the hyper heap size according to the memory size. We have extra expenses tracking memory we've mapped, this has to be accessible from R0. However, we don't really need this in RC, so it's one of the allocations we could put in a R3/R0 only heap later.
------------------------------------------------------------------------