commit_23128 ------------------------------------------------------------------------ r23128 | vboxsync | 2009-09-18 20:50:55 +0800 (Fri, 18 Sep 2009) | 1 line
COM/XPCOM glue: Bug fixes around the EventQueue init/uninit calls from com::Initialize and com::Shutdown. As com::Initialize() clearly states, this function will be called by all client threads doing COM/XPCOM work. Calling it more than once on the same thread is also documented as normal behavior (this is also evident from the counting done in the XPCOM version of the code). However it seems like the media checker is one of the few use cases for this and it doesn't kick in on all systems it seems... So, I hope this actually works as I cannot reproduce it here.
------------------------------------------------------------------------