commit_24090 ------------------------------------------------------------------------ r24090 | vboxsync | 2009-10-27 00:17:35 +0800 (Tue, 27 Oct 2009) | 2 lines
Main/Machine: when opening a new session, don't just fail of the previous session wasn't closed completely. Better wait a little, as normally it takes much less than half a second to clean it up. The client could retry, but usually that would be much slower as the client has no knowledge about when the session is closed.

------------------------------------------------------------------------