commit_24094 ------------------------------------------------------------------------ r24094 | vboxsync | 2009-10-27 05:39:34 +0800 (Tue, 27 Oct 2009) | 1 line
DevATA: Optimize the suspend and poweroff pendig I/O waits so they won't wait in 100ms chunks but block on the thread semaphore and get woken up by the I/O thread when it is idle. Combined the per-controller waiting with the wait for both controllers, sacrifying the timeout a bit in the latter case (but it's not all that important, so who cares). Made the constructor return on semaphore and thread creation failures. Moved the semaphore cleanup to the constructor and simplified the waiting there to just wait for the threads. Made the I/O thread stop setting its handle to NIL_RTTHREAD before quitting and causing a thread handle leak (IPRT internal handle).
------------------------------------------------------------------------