commit_24344 ------------------------------------------------------------------------ r24344 | vboxsync | 2009-11-05 00:34:16 +0800 (Thu, 05 Nov 2009) | 1 line
Main: fix a really nasty problem in the Machine destructor which never detached non-harddisk media when the machine object was deleted. The problem exhibited itself that when a machine was inaccessible for some reason (e.g. hard disk inaccessible) and also had non-hard disk media attached, debug code would assert somewhere very deep in media backreferences because first the machine was loaded, then marked inaccessible and destroyed (except for the non-harddisk media which never got detached), and then Machine::GetAccessible would retry loading that machine and couldn't attach anything because the old dead attachment was still in the backreferences
------------------------------------------------------------------------