commit_25073 ------------------------------------------------------------------------ r25073 | vboxsync | 2009-11-29 05:11:05 +0800 (Sun, 29 Nov 2009) | 3 lines
VMMDevHGCM.cpp: VBOXHGCMCMD::cbCmd isn't very helpful now that the rules of the saved state game has changed.  When loading / teleporting the state on a machine where VBOXHGCMCMD or/and VBOXHGCMSVCPARM is larger (either due to it being 64-bit or because of different alignment rules (GCPhys & MSC)), we would allocate too little memory for it and end up corrupting the heap.

As a very temporary measure, I have quadrupled cbSize before passing it to RTMemAllocZ and when used in comparisons.  (The value in VBOXHGCMCMD::cbCmd should not be changed in case it is saved again.)
------------------------------------------------------------------------