commit_25638 ------------------------------------------------------------------------ r25638 | vboxsync | 2010-01-05 00:08:04 +0800 (Tue, 05 Jan 2010) | 1 line
iprt,pdmcritsect: Added RTSemEvent[Set|Add|Remove]Signaller so that we can validate who is signalling an event if we like and, more importantly, detect deadlocks involving event semaphores.  More attempts at dealing with the races (and bugs) in the all-other-threads-blocking detection in tstRTLockValidator.cpp, adding RTThreadGetReallySleeping and RTThreadGetNativeState in the process.
------------------------------------------------------------------------