commit_26089 ------------------------------------------------------------------------ r26089 | vboxsync | 2010-01-29 01:08:19 +0800 (Fri, 29 Jan 2010) | 5 lines
Branding: Moved the Product & Vendor strings to kBuild, so it could be used
there as well. Added a Copyright year define which points to the current year.
All this should be used on more places. For now the help strings of the
Frontends and most strings of the Mac OS X installer are updated.

------------------------------------------------------------------------