commit_27138 ------------------------------------------------------------------------ r27138 | vboxsync | 2010-03-07 07:01:47 +0800 (Sun, 07 Mar 2010) | 1 line
Config.kmk,recompiler/Makefile.kmk,xpcom18a4/Makefile.kmk: Use @rpath and / in -rpath on darwin when not building the hardened version. This is more or less identical to what we do with$ORIGIN on linux and solaris. Will attempt to combine the two later.  Also, set the -compatibility_version to the same as -current_version to prevent accidentally mixing stuff from /Applications/VirtualBox.app/ with stuff from your dev build (and vice versa).
------------------------------------------------------------------------