commit_28405 ------------------------------------------------------------------------ r28405 | vboxsync | 2010-04-16 18:37:39 +0800 (Fri, 16 Apr 2010) | 2 lines
FE/SDL: When sending two key events to the guest (the first event was delayed to decide if this is part of a host key combination or not) add a small delay to work around a bug of certain guest applications (e.g. mstsc.exe on WinXP which otherwise wouldn't detect the first key event.

------------------------------------------------------------------------