commit_29036 ------------------------------------------------------------------------ r29036 | vboxsync | 2010-05-04 22:54:12 +0800 (Tue, 04 May 2010) | 2 lines
Main/Snapshot: handle late failures more gracefully when the actual merging succeeded but something else failed. If the merge source is gone assume the merge worked and update the snapshot accordingly. Avoids inaccessible VM configs due to referring to an already gone image. Small additional cleanup with handling the flag whether VirtualBox.xml needs to be saved - do not depend on getting the "OR" done by the called method, do it explicitly.

------------------------------------------------------------------------