commit_29932 ------------------------------------------------------------------------ r29932 | vboxsync | 2010-06-01 09:40:07 +0800 (Tue, 01 Jun 2010) | 1 line
Main: Rewrote ProgressProxyImpl.cpp to proxy at the most one progress object and skip the final operation of the proxy object that wasn't really needed in the openRemoteSession case. Operations are proxied as well so that the openRemoteSession user can wait for the teleporter to become operational by waiting for the 3rd last operation to complete in a loop(!).
------------------------------------------------------------------------