commit_33563 ------------------------------------------------------------------------ r33563 | vboxsync | 2010-10-28 22:46:26 +0800 (Thu, 28 Oct 2010) | 1 line
iprt:ministring: Added the java-style equals() and equalsIgnoreCase() as equals() can optimize the comparison by first checking if the length is the same (compare() cannot as it needs to determin the ordering).  Added appendCodePoint() for UTF-8.  Fixed the incorrect assumption in toUpper and toLower that the string length remained unchanged - the string might shrink as the folded code points may have a shorter encoding. Added testcase that verifies that a code point will not grow during folding and that have a stable encoding length after it has been changed in a folding.
------------------------------------------------------------------------