commit_34326 ------------------------------------------------------------------------ r34326 | vboxsync | 2010-11-24 22:03:55 +0800 (Wed, 24 Nov 2010) | 1 line
VMM: Removed the XXXInitCPU and XXXTermCPU methods since all but the HWACCM ones where stubs and the XXXTermCPU bits was not called in all expected paths. The HWACCMR3InitCPU was hooked up as a VMINITCOMPLETED_RING3 hook, essentially leaving it's position in the order of things unchanged, while the HWACCMR3TermCPU call was made static without changing its position at the end of HWACCMR3Term.
------------------------------------------------------------------------