commit_35757 ------------------------------------------------------------------------ r35757 | vboxsync | 2011-01-28 20:51:37 +0800 (Fri, 28 Jan 2011) | 1 line
Main: do not fail loading machine settings if a shared folder path is not absolute (that breaks importing OVF and machine folders from other hosts since the rules about what constitutes an absolute path differ between host OSes). Instead, issue a runtime warning if a shared folder path is not absolute or does not exist on the host when the VM is powered up.
------------------------------------------------------------------------