commit_36401 ------------------------------------------------------------------------ r36401 | vboxsync | 2011-03-24 21:16:30 +0800 (Thu, 24 Mar 2011) | 1 line
VBoxBFE: The testcase/Makefile.kmk is *NOT* allowed to reference variables defined by the parent makefile, that *BREAKS* running make in the testcase subdirectory. tstMouseImpl must use the INTERMEDIATES property just like VBoxBFE to make sure the headers are generated before sources are built, that's probably what caused the build break that r70348 and others tried to fix.
------------------------------------------------------------------------