commit_36965 ------------------------------------------------------------------------ r36965 | vboxsync | 2011-05-05 11:03:22 +0800 (Thu, 05 May 2011) | 9 lines
NAT: Attempt to fix (the rest of vbox/7711 and vbox/8701).
The reason of such corruption is aliasing of fragments, libalias
processing fragments diffenrently from first fragment,
it's set aliasing address to the source IP address. That insults
guest TCP/IP stack and it drops the rest of the fragments. In this
changeset we let chewing first packet, letting libalias apply its
modules changes if any and do not pass the rest of fragments, because
libalias can't recognize the fragments anyway.

------------------------------------------------------------------------