commit_38717 ------------------------------------------------------------------------ r38717 | vboxsync | 2011-09-12 22:48:50 +0800 (Mon, 12 Sep 2011) | 2 lines
Main/Medium+Machine+AutoLock: Integrate the queryInfo semaphore into the AutoLock infrastructure with the appropriate lock class. Adjust the locking rules for queryInfo and start propagating them up to the callers. Needs more work, but this is comparably simple: just get the mediaTree lock in the caller if the assert triggers.

------------------------------------------------------------------------