commit_40491 ------------------------------------------------------------------------ r40491 | vboxsync | 2012-03-16 04:52:05 +0800 (Fri, 16 Mar 2012) | 2 lines
Main/NetworkAdapter: Take the easy solution out of the reference hell created by the SessionMachine logic. Since the BandwidthGroup objects are not logically located below NetworkAdapter (they are under Machine/BandwidthControl) it is effectively impossible to get the right references in the object. Querying through the appropriate Machine object is guaranteed to give the right result.

------------------------------------------------------------------------