commit_42892 ------------------------------------------------------------------------ r42892 | vboxsync | 2012-08-21 04:48:06 +0800 (Tue, 21 Aug 2012) | 1 line
BlockCache: Relax check when restoring a state. It is allowed to have more cache users in the saved state than there are registered currently if the missig users don't have any data saved. Should fix restore errors if a VM with snapshots was moved to a filesystem with known data corruption bugs which disables async I/O whe the VM starts
------------------------------------------------------------------------