commit_42958 ------------------------------------------------------------------------ r42958 | vboxsync | 2012-08-24 02:17:28 +0800 (Fri, 24 Aug 2012) | 1 line
ATAPI/Passthrough: Don't try to detect the medium type if a READ TOC/PMA/ATIP command with the CD-Text format is sent and reduce the number of required bytes of the reply to 4. This is enough to detect the data type on a medium. Fixes audio CD passthrough with certain media players like vlc, kaffeine and amarok
------------------------------------------------------------------------