commit_42987 ------------------------------------------------------------------------ r42987 | vboxsync | 2012-08-25 09:36:58 +0800 (Sat, 25 Aug 2012) | 1 line
Installer/darwin: drop checkforrunningvm in favor of a one-liner. That should work on Leopard too. Dropped the InstallationCheck in VBoxKEXTs which does the same as the javascript in distribution.dist already did and may cause later install errors with generic message (probably because we didn't ship InstallationCheck.strings, but whatever).
------------------------------------------------------------------------