commit_44305 ------------------------------------------------------------------------ r44305 | vboxsync | 2013-01-16 20:01:19 +0800 (Wed, 16 Jan 2013) | 6 lines
VboxTray/VideoMode: Dynamic configuration of secondary monitor for windows guest. (Enabling/disabling/Change of resolution/position).

Fix for the error reported on #6118 comment 46, "setvideomodehint 1024*768 32 1 yes" not working.
If the previous resolution of the disabled secondary screen is 1024*768*32 and command is issued"setvideomodehint 1024*768 32 1 yes", then there is a conditional block in the code returns TRUE assuming that it is already at desired resolution and no further action is required.

To fix the issue, now I am checking that if the monitor, for which the request is being made,  is disabled then conditional block will return FALSE and code will proceed further to enable the monitor.
------------------------------------------------------------------------