commit_44382 ------------------------------------------------------------------------ r44382 | vboxsync | 2013-01-25 22:14:51 +0800 (Fri, 25 Jan 2013) | 3 lines
Main/DisplayImpl: Fix for access violation in case of multimonitor.

Display requests a resize of 4 framebuffers simultaneously. When one of framebuffers is resized, Display calls InvalidateAndUpdateEMT which is updating _all_ framebuffers but some of them are not yet finished with the resize. As a result, InvalidateAndUpdateEMT is assuming that all the framebuffers have right size, which is not true. Thus, crash while copying from source to destination recatangle .
------------------------------------------------------------------------