commit_44668 ------------------------------------------------------------------------ r44668 | vboxsync | 2013-02-13 22:53:36 +0800 (Wed, 13 Feb 2013) | 1 line
DevCodec.*: More cleanups... moving all the unnecesary defines from the header to the implementation. Use pThis instead of pState and pCodecState, with a couple of exceptions. Assert() does NOT need double parentheses. Just AssertFailedReturn() if you cannot resolve the input into a register (or whatever), shorter+faster than initializing the register point to NULL and check if it's still NULL after the resolver code.
------------------------------------------------------------------------