commit_45676 ------------------------------------------------------------------------ r45676 | vboxsync | 2013-04-23 18:26:14 +0800 (Tue, 23 Apr 2013) | 3 lines
VBoxTray/VBoxSeamless: Fix for the issue - A small rectangular bar appears over the start button for Vista and above in seamless mode when mouse is hovered over the start button..
Solution:
There is difference in the rectangle size of "Start" button and its shadow as reported by the API GetWindowRect. So , adjusting the size of the "Start" button to match the shadow.
------------------------------------------------------------------------