commit_45845 ------------------------------------------------------------------------ r45845 | vboxsync | 2013-05-01 00:38:34 +0800 (Wed, 01 May 2013) | 5 lines
VMM/VMMR0: Load kernel-gs base MSR as part of the auto-load/store MSR feature.
VMM/HMVMXR0: Fixed the code to also deal with the case when auto-load/store is not defined.
VMM/VMMR0: Fixed LegacyandAmd64.mac to restore the MSRs from the stack in the right order, not sure how it
could have worked previously.

------------------------------------------------------------------------