commit_46065 ------------------------------------------------------------------------ r46065 | vboxsync | 2013-05-14 21:15:42 +0800 (Tue, 14 May 2013) | 3 lines
Main/DisplayImpl: 6711: Fix for the race condition where in GA sends the new resize rectangle dimensions while framebuffer is still resizing. At this moment DisplayImpl doesn't have the updated framebuffer rectangles and as a result the intersection between the new rectangles passed by GA and current frambuffe is NULL. Thus DisplayImpl rejects the rectangles passed from GA and GUI receives NULL for drawing.
The fix for the issue is to let DisplayImpl save the rectangles passed from the GA till the point framebuffer is completely resized and then pass these rectangle to GUI after checking for intersection with the curent newly resized framebuffer.

------------------------------------------------------------------------