commit_46494 ------------------------------------------------------------------------ r46494 | vboxsync | 2013-06-11 23:03:22 +0800 (Tue, 11 Jun 2013) | 1 line
ipcdclient.cpp: Use a shared critical section instead of the slowlish PRMonitor thingy. The origianl code mentions that a monitor is a bit overkill and PRLock would've sufficed if it only was recursive. I've tried taking this a bit further by using a shared lock so that GetTarget() and IPC_OnMessageAvailable() doesn't need to do exclusive access.
------------------------------------------------------------------------