commit_47736 ------------------------------------------------------------------------ r47736 | vboxsync | 2013-08-15 03:49:47 +0800 (Thu, 15 Aug 2013) | 1 line
HNVMXR0.cpp: Separate run-loop for single stepping to catch exceptions and such without harming performance of the normal execution (requires RIP and CS to be saved). Simplified the UNUSABLE handling in hmR0VmxReadSegmentReg to just make sure P and a few irrelevant other bits are cleared. Always clearing the G would cause the limit to change on reload. If having bits set here is causing trouble, we'll fix the other code making the trouble, not work around it here.
------------------------------------------------------------------------