commit_48178 ------------------------------------------------------------------------ r48178 | vboxsync | 2013-08-30 11:54:49 +0800 (Fri, 30 Aug 2013) | 9 lines
Do not hardcode lwip location in lwip makefile.  Provide functions for
"calling" makefiles to add lwip to its sources and use these functions
in Devices and NAT.

XXX: It's still icky that Devices pulls in lwip makefile and then NAT
(which is included later) inherits it.  But this commit allows
inclusion problem (and possible move of the lwip directory) to be
addressed separately.

------------------------------------------------------------------------