commit_48779 ------------------------------------------------------------------------ r48779 | vboxsync | 2013-10-01 10:14:13 +0800 (Tue, 01 Oct 2013) | 1 line
iprt/xml.h: Replaced std::list<shared_ptr<>> with RTList both to safe a bunch of heap allocations but more importantly to improve enumerability of the XML tree. Added some new methods too, both related to enumeration and to other things. Note: If we can get rid of the shared_ptr stuff with attributes we can jettison the boost dependency we have, AFAICT.
------------------------------------------------------------------------