commit_48810 ------------------------------------------------------------------------ r48810 | vboxsync | 2013-10-02 16:48:34 +0800 (Wed, 02 Oct 2013) | 2 lines
Linux installers: Move the udev rule execution order from 10 to 60. The reason for having it at 10 was to allow other rules to overwrite these rules, mainly permissions of the USB devices. This is not longer necessary (we are using /dev/vboxusb for a long time now) and we need to take care that certain systems like EL6 don't reset the permissions of /dev/vboxdrvu -- the unrestricted device -- back to 0600 (we want 0666 here).

------------------------------------------------------------------------