commit_49102 ------------------------------------------------------------------------ r49102 | vboxsync | 2013-10-15 12:18:36 +0800 (Tue, 15 Oct 2013) | 8 lines
Listen for VBoxEventType_OnNATNetworkSetting to pick up changes in
AdvertiseDefaultIPv6RouteEnabled.

XXX: We also need to prod rtadvd to send unsolicited advertisement
with updated router lifetime, but event delivery seems to be busted,
so I can't test that properly.  For now just pick up the flag, that
will be announced on the next scheduled advertisement.

------------------------------------------------------------------------