commit_49854 ------------------------------------------------------------------------ r49854 | vboxsync | 2013-12-10 04:52:21 +0800 (Tue, 10 Dec 2013) | 7 lines
"IP_HDRINCL" is a name of a socket option and the #undef of it is a
good sign that reusing this name was rather unfortuate.  Other naming
conflicts are mostly restricted to "lwip/sockets.h" but this one is
pretty nasty since it affects all code that uses lwip.  Rename it out
of the way in "lwip/ip.h" and do undef/redef dance in the few lwip
files that use this name to minimize diffs to upstream.

------------------------------------------------------------------------