commit_50074 ------------------------------------------------------------------------ r50074 | vboxsync | 2014-01-14 13:14:14 +0800 (Tue, 14 Jan 2014) | 10 lines
The meaning of LWIP_PREFIX_BYTEORDER_FUNCS was inverted in lwip-1.4
and that was not accounted for when old lwipopts.h was used as the
base for the new one.

Undef LWIP_PREFIX_BYTEORDER_FUNCS in lwipopts.h and drop the code that
had to be added to lwip-namespace.h to compensate - it does the same
redefinition dance that lwip's def.h already does.

Same object code is generated.

------------------------------------------------------------------------