commit_50949 ------------------------------------------------------------------------ r50949 | vboxsync | 2014-04-02 08:53:08 +0800 (Wed, 02 Apr 2014) | 6 lines
NAT: Rename horrifying "VBOX_NAT_HNCE_EXSPOSED_NAME_RESOLUTION_INFO"
which is not just a mouthful, but has a typo in "exposed" too.  While
here rename the rest of them with s/HNCE/DNS/.  "HNCE" is an obscure
abbreviation and as far as I can tell it doesn't really reflect the
meaning of these constants.

------------------------------------------------------------------------