commit_50950 ------------------------------------------------------------------------ r50950 | vboxsync | 2014-04-02 09:32:02 +0800 (Wed, 02 Apr 2014) | 8 lines
NAT: Rename fHostNetworkConfigurationEventListener parameter to fFlapLink.
drvNATHostNetworkConfigurationChangeEventStrategySelector is called
with it set to true not only for the (MacOS) event.  I can see how it
is intended to eventually signal the fact that DNS information *has*
changed, but then it wouldn't make sense to do anything if it hadn't.
So use the name that better describes current reallity and deal with
the future in the future.

------------------------------------------------------------------------