commit_51908 ------------------------------------------------------------------------ r51908 | vboxsync | 2014-07-08 01:44:56 +0800 (Tue, 08 Jul 2014) | 1 line
VBoxNetFlt/win: Unbreak VBoxNetFltUninstall - apparently the previous changes made VBoxNetCfgWinNetFltUninstall return S_FALSE when we're not installed. Dunno if it can return this in other cases. Undid some of the Id vs ID inconsistencies introduced by that same changeset - Id was used before and that's the preferred way of doing it.
------------------------------------------------------------------------