commit_51974 ------------------------------------------------------------------------ r51974 | vboxsync | 2014-07-10 22:45:05 +0800 (Thu, 10 Jul 2014) | 8 lines
NAT/Net: Careful with that DWORD... It's unsigned, so assigning
negative values to it is a bad idea, especially when a coercion to a
wider signed type is about to happen.  While here, don't test that
DWORD values are <= 0, just test for zero instead.

In particulat this fixes returning socket errors from
pxtcp_sock_recv() and pxtcp_sock_send() wrappers on Windows.

------------------------------------------------------------------------