commit_52394 ------------------------------------------------------------------------ r52394 | vboxsync | 2014-08-16 06:25:31 +0800 (Sat, 16 Aug 2014) | 1 line
intnetR0TrunkIfPortDisconnect: Must always release the busy reference to the interface side (netflt) and must not trust pThis->pIfPort to be working as it's cleared before doing the busy waiting in intnetR0TrunkIfDestroy. So, pfnDisconnect has been extended with a pIfPort pointer as well as a function pointer to a busy-release function for pIfPort.
------------------------------------------------------------------------