commit_52417 ------------------------------------------------------------------------ r52417 | vboxsync | 2014-08-19 23:11:12 +0800 (Tue, 19 Aug 2014) | 1 line
Startup optimization: Don't check the hardened-program.dll signature before we need it, because we won't be loading this DLL in the first two processes we spawn anyway and it's a waste of time to check a huge DLL two times extra. We'll check it when loading the DLL, so no worries.
------------------------------------------------------------------------